+++ main.py
@@ -1,9 +1,11 @@
-import os
 import logging
 
 from telegram.ext import CommandHandler
 from telegram.ext import Updater
 from telegram.ext import MessageHandler, Filters
+
+from jobs import check_server
+from tools import get_token
 
 
 # setting up logging so that we know if things are breaking
@@ -14,13 +16,6 @@ logging.basicConfig(
 )
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
-
-
-def get_token():
-    path = os.path.dirname(os.path.abspath(__file__))
-    token_path = os.path.join(path, 'token.txt')
-    token = open(token_path, 'r').read().rstrip()
-    return token
 
 
 def start(bot, update):
@@ -38,16 +33,6 @@ def echo(bot, update):
         text=update.message.text,
     )
 
-def check_server(bot, job):
-    chat_id, hosts = job.context[0], job.context[1] 
-    for hostname in hosts:
-        response = os.system("ping -c 1 " + hostname)
-        if response == 0:
-            bot.send_message(
-                chat_id=chat_id,
-                text=f"{hostname} is up",
-            )
-
 
 def start_monitoring(bot, update, job_queue):
     logger.info('Start monitoring was called')
@@ -56,9 +41,9 @@ def start_monitoring(bot, update, job_queue):
     message_text.remove('')
     context = [update.message.chat_id, message_text]
     job_queue.run_repeating(
-        check_server, 
-        interval=3, 
-        first=0, 
+        check_server,
+        interval=3,
+        first=0,
         context=context,
     )
 
@@ -74,10 +59,11 @@ if __name__ == '__main__':
     dispatcher = updater.dispatcher
 
     start_handler = CommandHandler('start', start)
-    start_monitoring_handler = CommandHandler('start_monitoring', 
-        start_monitoring, 
+    start_monitoring_handler = CommandHandler(
+        'start_monitoring',
+        start_monitoring,
         pass_job_queue=True,
-    )     
+    )
 
     echo_handler = MessageHandler(Filters.text, echo)
